requirements.txt
langchain langchain-openai
langchain-postgres[psycopg]
langchain-mongodb
langchain-oracledb
mysql-connector-python
oracledb 
pymongo
gradio